refactor(cli): split package command dispatch
